+++ Changelog.md
@@ -2,11 +2,10 @@
 
 {:toc}
 
-## Changelog of CUDA
+## Changelog of Deformable-DETR
 
 ### 1.0.0 @ 01/13/2025
 
 #### :mega: New
 
-1. Create this project.
-2. Finish the script of building `CUDA 12.x` on `Debian`.
+1. Finish the script of building `Deformable-DETR` on `Python:3.12-slim` and `CUDA:12.4`.